# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fikrie Portfolio API
  description: |
    REST API for the Fikrie Portfolio Next.js application.

    ## Authentication
    Most write endpoints require a session cookie:
    - Cookie name: `portfolio_session`
    - Cookie value: `authenticated`

    Log in via the `/login` page (Server Action, not a REST endpoint) to obtain the cookie,
    then include it in subsequent requests.

    ## Data storage
    - JSON data: `data/*.json`
    - Uploaded files: `public/` subdirectories
  version: 1.0.0
  contact:
    name: Fahreza Portfolio
servers:
  - url: http://localhost:3000
    description: Local development
  - url: http://localhost:3001
    description: Playwright / alternate dev port
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: portfolio_session
      description: Set to `authenticated` after logging in
  schemas:
...
    Review:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        company: { type: string }
        content: { type: string }
        image: { type: string }
      required: [id, name, role, company, content, image]
...
paths:
...
  /api/settings/reviews:
    get:
      tags: [Settings]
      summary: List client reviews
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items: { $ref: '#/components/schemas/Review' }
    post:
      tags: [Settings]
      summary: Add a client review
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2 }
                role: { type: string }
                company: { type: string }
                content: { type: string, minLength: 10 }
                image: { type: string, format: binary }
                imageUrl: { type: string }
              required: [name, content]
      responses:
        '200': { description: Returns review and reviews array }
        '401': { description: Unauthorized }
    patch:
      tags: [Settings]
      summary: Update a client review
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                reviewId: { type: string }
                name: { type: string }
                role: { type: string }
                company: { type: string }
                content: { type: string }
                image: { type: string, format: binary }
                imageUrl: { type: string }
              required: [reviewId]
      responses:
        '200': { description: Success }
        '401': { description: Unauthorized }
    delete:
      tags: [Settings]
      summary: Delete a client review
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewId: { type: string }
              required: [reviewId]
      responses:
        '200': { description: Returns updated reviews array }
        '401': { description: Unauthorized }